+++ src/android/FirebasePlugin.java
@@ -259,6 +259,8 @@ public class FirebasePlugin extends CordovaPlugin {
 
             } else if (action.equals("verifyPhoneNumber")) {
                 this.verifyPhoneNumber(callbackContext, args);
+            } else if (action.equals("setLanguageCode")) {
+                this.setLanguageCode(callbackContext, args);
             } else if (action.equals("authenticateUserWithGoogle")) {
                 this.authenticateUserWithGoogle(callbackContext, args);
             } else if (action.equals("authenticateUserWithApple")) {
@@ -1456,6 +1458,27 @@ public class FirebasePlugin extends CordovaPlugin {
         });
     }
 
+    public void setLanguageCode(final CallbackContext callbackContext, final JSONArray args){
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    String lang = args.getString(0);
+
+                    if(lang == null || lang.equals("")){
+                        callbackContext.error("Lang must be specified");
+                        return;
+                    }
+
+                    FirebaseAuth.getInstance().setLanguageCode(lang);
+
+                    Log.d(TAG, "Language code setted to "+lang);
+                } catch (Exception e) {
+                    handleExceptionWithContext(e, callbackContext);
+                }
+            }
+        });
+    }
+    
     public void createUserWithEmailAndPassword(final CallbackContext callbackContext, final JSONArray args){
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {