+++ src/ios/FirebasePlugin.m
@@ -543,6 +543,16 @@ static NSMutableDictionary* firestoreListeners;
     }
 }
 
+- (void)setLanguageCode:(CDVInvokedUrlCommand *)command {
+    NSString* lang = [command.arguments objectAtIndex:0];
+    @try {
+         [FIRAuth auth].languageCode = lang;
+         NSLog(@"Language code setted to %@!", lang);
+    }@catch (NSException *exception) {
+        [self handlePluginExceptionWithContext:exception :command];
+    }
+}
+
 - (void)createUserWithEmailAndPassword:(CDVInvokedUrlCommand*)command {
     @try {
         NSString* email = [command.arguments objectAtIndex:0];