+++ CHANGELOG.md
@@ -3,6 +3,8 @@
     * Bug introduced by PR [#482](https://github.com/dpa99c/cordova-plugin-firebasex/pull/482).
 * (Android) Add the Firebase Performance Monitoring Gradle plugin to monitor network traffic.
     * Resolves [#520](https://github.com/dpa99c/cordova-plugin-firebasex/issues/520).
+* (Feature): Add setLanguageCode method for Firebase Auth
+    * Merged from PR [#527](https://github.com/dpa99c/cordova-plugin-firebasex/pull/527).
 
 # Version 11.0.1
 * (iOS) Set the Sign In with Apple capability based on the `IOS_ENABLE_APPLE_SIGNIN` plugin variable.